Reduce code duplication between EventsFixture shared tests and the one that is specific only to NUnit framework.
